Rename documentinverter* -> document_inverter*.
