+++ searchlib/CMakeLists.txt
@@ -175,7 +175,7 @@ vespa_define_module(
     src/tests/ld-library-path
     src/tests/memoryindex/compact_document_words_store
     src/tests/memoryindex/datastore
-    src/tests/memoryindex/documentinverter
+    src/tests/memoryindex/document_inverter
     src/tests/memoryindex/field_index
     src/tests/memoryindex/field_index_remover
     src/tests/memoryindex/fieldinverter
