+++ searchlib/src/tests/memoryindex/document_inverter/.gitignore
@@ -1,0 +1,1 @@
+searchlib_document_inverter_test_app
