+++ searchlib/src/tests/memoryindex/document_inverter/document_inverter_test.cpp
@@ -3,7 +3,7 @@
 
 
 #include <vespa/searchlib/index/docbuilder.h>
-#include <vespa/searchlib/memoryindex/documentinverter.h>
+#include <vespa/searchlib/memoryindex/document_inverter.h>
 #include <vespa/searchlib/memoryindex/fieldinverter.h>
 #include <vespa/searchlib/test/memoryindex/ordered_field_index_inserter.h>
 #include <vespa/searchlib/common/sequencedtaskexecutor.h>
