--- searchlib/src/tests/memoryindex/documentinverter/.gitignore
@@ -1,1 +1,0 @@
-searchlib_documentinverter_test_app
